Add router to select character and show backgroung color on selected item
