+++ package.json
@@ -11,7 +11,8 @@
   "dependencies": {
     "react": "^18",
     "react-dom": "^18",
-    "next": "13.5.11"
+    "next": "13.5.11",
+    "react-router-dom": "^6.14.1"
   },
   "devDependencies": {
     "typescript": "^5",
